Move tests into slow file
