+++ ironfish/src/blockchain/__fixtures__/blockchain.test.slow.ts.fixture
@@ -1,0 +1,82 @@
+{
+  "Blockchain asset updates with a mint description upserts an asset to the database": [
+    {
+      "id": "41942454-6111-44d6-88e5-ed4a87562ffb",
+      "name": "test",
+      "spendingKey": "f095939827f874714e432cbfe22dbf7374dd25009fdac8f4ee15d71909e9ab14",
+      "incomingViewKey": "c3b9a9f15b57a9e8a4d40df4a041d51563f7bd49d03f0926394f8f9cc7b11c05",
+      "outgoingViewKey": "65389c646c678d7afb1b569393cf8e46b936fb78e5436a32c4fb22a7b2509a3c",
+      "publicAddress": "4b29c1c491ef09507ebee7a64df2f7526f18578a627713a6e6b1c7b5b439c680"
+    }
+  ],
+  "Blockchain asset updates with a burn description decrements the asset supply from the database": [
+    {
+      "id": "b730350e-8321-4262-84a4-36cf08818a5b",
+      "name": "test",
+      "spendingKey": "e31579707097f309d285d95c134f7d7838f1011d3c9fc7833190f8f303fe2ec6",
+      "incomingViewKey": "d03e8f3af5433d3967c994e342ad26f024b83a5ba593ee3cb1a19aed8cf4a303",
+      "outgoingViewKey": "7ca1565d4a845e51ca7bc98b939e8242c44ec663ae9ce586f87288b605056e35",
+      "publicAddress": "1da49e8236c7efe4122bd085eee02877b3625e6c3f9bc5f06f0c84b392d824bc"
+    }
+  ],
+  "Blockchain asset updates with a subsequent mint should keep the same created transaction hash and increase the supply": [
+    {
+      "id": "6b687b2f-6834-42b9-ac47-cb3bfe8d0160",
+      "name": "test",
+      "spendingKey": "0afba21829fc46747a08a460f85eb05fa717630f87b9c9241618e1d9be1b0a7d",
+      "incomingViewKey": "25635ce2bb3bbe73dbb5a94f3a713246cb60e7c2a93f4d58078aa37e54fdb802",
+      "outgoingViewKey": "057afa75e14274899414a43d6a1ecac755c591fdb47258e75d468d2e29ff54fb",
+      "publicAddress": "94b1d132555e449bd08c5ed0893cd07157cff0cc16863ddc29b0e8e399df73d4"
+    }
+  ],
+  "Blockchain asset updates when the first mint gets rolled back should delete the asset": [
+    {
+      "id": "678ed132-bbee-455a-83c1-44b3b8d78366",
+      "name": "test",
+      "spendingKey": "c1b580ea2821b6d1d56e720ba3724314546ebd00fe06b71661532e65e12f1fe8",
+      "incomingViewKey": "7827242ddca5a2331f84a4023fadaeeca1ac335e9c0c9ce337372f37de054c03",
+      "outgoingViewKey": "25cb9ff85182c6356b14de68f2ca5d8d0e189033986ce82a405b7d2df1ee305e",
+      "publicAddress": "e10a460197ba62be659be0b55b1112e3cc206d0cd3579cfc2f67c5dd70cdc09f"
+    }
+  ],
+  "Blockchain asset updates when a subsequent mint gets rolled back should decrement the supply": [
+    {
+      "id": "6e8cbed6-dafe-4c1c-b898-45fa4f26ec3f",
+      "name": "test",
+      "spendingKey": "1ecdf589ccad86dcae24cab39c559229854006a4631668ad5a0e9c9e4f2a84d0",
+      "incomingViewKey": "517a1bc81a3870f0495bd3984f6adc486174d23a7b19f68ddf996a2499657d07",
+      "outgoingViewKey": "2f93aca12d95a5808b55c270ea93879426148ccfdcc834b2461534766762fd04",
+      "publicAddress": "c8bedaba60946e76c5c1fad61b6ab89c8225ddcdcb194fecb20f3ab3d298b667"
+    }
+  ],
+  "Blockchain asset updates when a burn gets rolled back should increase the supply": [
+    {
+      "id": "e4968a9d-88fd-4866-99d6-a0bdd65eecae",
+      "name": "test",
+      "spendingKey": "fb363dc29e2597d939c0217fc29368795d1631f95d10e6b093ffee3f7bbd5548",
+      "incomingViewKey": "63d9e3eeba8dc7a26c0ff41362aff3018e385360fad3be8625861416e06a9f07",
+      "outgoingViewKey": "fb8011b41fb3a40617d10ac62cc0c1a61ae05fd9b0c6313901045067e72ff8b8",
+      "publicAddress": "fa2680e09982568f9755d142351e2eaf5fcfbe736d2370698053f2feb2f4e186"
+    }
+  ],
+  "Blockchain asset updates when burning too much value throws an exception": [
+    {
+      "id": "bd6f0333-93a8-4db6-aa73-880dce17e2bf",
+      "name": "test",
+      "spendingKey": "a5756a0d44cebf2f967765c18b6411315ae33d962c2706f29505ee4e0806b1db",
+      "incomingViewKey": "e52c36c6cfba2335a80cc7b1b5319b3bba455f42715dd9d69091ec9e48aac500",
+      "outgoingViewKey": "21c331dfc9f8f2cc606572be14d66ad8289153398ebe00d01fdbdcde3cc821f0",
+      "publicAddress": "3e221fb6879291c68b54fd55a05c4f633f648f59daa44ec06430567f198e2759"
+    }
+  ],
+  "Blockchain asset updates when rolling back multiple mints and burns adjusts the supply accordingly": [
+    {
+      "id": "6512589c-ca53-41f0-a91c-d8c79dc7a2ea",
+      "name": "test",
+      "spendingKey": "260e0d61a4f6c5ad3c28af2e17155d7f004b1027f1fad8ffe5a70082a832aba1",
+      "incomingViewKey": "7ff6369f57d0dda9a3880f39a7b059065faacb709c6484beffe1024c2ba4da07",
+      "outgoingViewKey": "6dcfa638a2fb560b784762ac52bbd1c8a304efdf06ce9931b534987dc9506b19",
+      "publicAddress": "d0fc5aae3176950849fa6cc3948a8a384fc6be9ccc6a42dd865d4c44a6c59bbc"
+    }
+  ]
+}
